feat: all keyboard keys are set; follow to functionality
