+++ css/style.css
@@ -1,6 +1,7 @@
 .page {
   margin: 0 auto;
   font-family: Helvetica, sans-serif;
+  background: #ccc;
 }
 
 .wrapper {
@@ -10,7 +11,6 @@
   box-sizing: content-box;
   min-height: 100%;
   overflow: hidden;
-  background: #ccc;
 }
 
 .description,
@@ -160,6 +160,7 @@
 
 .keyboard .Enter {
   height: 85px;
+  /* clip-path: polygon(35% 0, 76% 0, 75% 100%, 44% 100%, 44% 36%, 35% 36%); */
 }
 
 .keyboard .ShiftLeft {
@@ -179,8 +180,9 @@
 .keyboard .Space {
   width: 330px;
 }
+
 .keyboard .Fn,
-.keyboard .CtrlLeft{
+.keyboard .CtrlLeft {
   display: flex;
   justify-content: flex-start;
   align-items: flex-start;
@@ -199,4 +201,3 @@
   height: 20px;
   font-size: 7px;
 }
-
